==> 02_cleaning.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)

df = pd.read_csv(r"SBS_Processed_Datasets/01_structured_output.csv",low_memory=False)

df.head(2)

df.columns

#Remove unwanted columns
unwanted_columns = [
    'new_car_feature', 'car_links', 'Insurance Validity', 'it', 'ft', 'bt', 'km', 'owner',
    'centralVariantId','priceActual', 'priceSaving', 'priceFixedText', 'trendingText', 'Seats', 
    'RTO', 'Max Power', 'Torque', 'Color', 'Engine Type', 'Displacement',
    'Max Torque', 'No of Cylinder', 'Values per Cylinder', 
    'Value Configuration', 'Fuel Suppy System', 'BoreX Stroke', 'Compression Ratio', 
    'Turbo Charger', 'Super Charger', 'Length', 'Width', 'Height', 'Wheel Base', 
    'Front Tread', 'Rear Tread', 'Kerb Weight', 'Gross Weight', 'Gear Box', 
    'Seating Capacity', 'Steering Type', 'Turning Radius', 'Front Brake Type', 
    'Rear Brake Type', 'Top Speed', 'Acceleration', 'Tyre Type', 'No Door Numbers', 
    'Cargo Volumn', 'Wheel Size', 'Alloy Wheel Size', 'Ground Clearance Unladen',
    'Drive Type','Ownership','Year of Manufacture','Transmission', 'Engine Displacement','Seats.1'
]

df = df.drop(columns=unwanted_columns)

df.columns

# Rename_columns
df.rename(columns={"ownerNo": "owner_no",
                   "oem": "brand",
                   "modelYear": "model_year",
                   "variantName": "variant_name",
                   "Registration Year": "registered_year",
                   "Fuel Type": "fuel_type",
                   "Kms Driven": "kms_driven",
                   "Engine": "engine_cc",
                   "Year of Manufacture": "manufacture_year",
                   "Mileage": "mileage_kmpl"},inplace=True)

# Strip_spaces_in_column_names
df.columns = df.columns.str.strip()

# Strip all white/trailing spaces & if string convert it to lowercase for consistency
df = df.apply(lambda col: col.str.strip().str.lower() if col.dtype == 'object' else col,axis=0)

df.head()

def convert_price_to_numerical_value(x):
    price = x.replace("₹","").replace(",","").lower().strip()

    if "thousand" in price or "thousands" in price:
        return float(price.replace("thousand" , "").replace("thousands", "").strip()) * 1_000
    elif "lakh" in price or "lakhs" in price:
        return float(price.replace("lakh","").replace("lakhs","").strip()) * 10_0000
    elif "crore" in price or "crores" in price:
        return float(price.replace("crore","").replace("crores","").strip()) * 10_00_000
    else:
        return float(price)

# remove 'kms' and make it numerical standard
df["kms_driven"] = df["kms_driven"].replace({"kms":"" , ",":""},regex=True).str.strip()

# engine displacement column remove unwanted strings
df["engine_cc"] = df["engine_cc"].str.replace("cc","").str.strip().astype(float)

# remove "kmph" in mileage_kmph column
df["mileage_kmpl"] = df["mileage_kmpl"].replace({"kmpl":"" , "km/kg": ""}, regex=True).str.strip().astype(float)

# remove unwanted strings and keep only year
df["registered_year"] = df["registered_year"].str.extract(r'\b(\d{4})\b').astype(float) #float will handle Nan value

# convert price colums values to its numerical format
df["price"] = df["price"].apply(lambda x: convert_price_to_numerical_value(x))
	


#drop duplicate records
print(df.duplicated().sum())  # Count duplicates
df = df.drop_duplicates()  # Drop duplicates

df.shape

df.info()

df.isnull().sum()

# Fill missing values in `registered_year` with corresponding `model_year`
df['registered_year'] = df['registered_year'].fillna(df['model_year'])

df.info()

df[df["engine_cc"].isnull()]

#both mileage and engine_cc are null so drop these rows
df.dropna(subset=['engine_cc'],inplace=True)

df.info()

df.isnull().sum()

# fill mileage_kmpl using imputation
df['mileage_kmpl'] = df.groupby(['model','engine_cc','fuel_type'])['mileage_kmpl'].transform(lambda x: x.fillna(x.mean()))

df.isnull().sum()

# Drop rows where `kms_driven` or `mileage_kmpl` has NaN values
df.dropna(subset=['kms_driven', 'mileage_kmpl'], inplace=True)

# convert to numeric datatypes
df["engine_cc"] = pd.to_numeric(df["engine_cc"],errors="coerce").fillna(0).astype(int)
df["registered_year"] = df["registered_year"].fillna(0).astype(int)
df["kms_driven"] = pd.to_numeric(df["kms_driven"],errors="coerce").fillna(0).astype(float)
df['mileage_kmpl'] = pd.to_numeric(df['mileage_kmpl'], errors='coerce').fillna(0).astype(float)
df['city'] = df['city'].astype('category')

df.head()

df.isnull().sum()

df.info()

df.describe()

df = df[(df["engine_cc"] >= 500)]

df.describe()

# Find the count of zero values in each column
zero_counts = (df == 0).sum()
print(zero_counts)

# Replace 0 with 1 in `owner_no`
df['owner_no'] = df['owner_no'].replace(0, 1)

df.shape

# Find the count of zero values in each column
zero_counts = (df == 0).sum()
print(zero_counts)

df["fuel_type"].unique()

df.head()

#Export it
df.to_csv(r"SBS_Processed_Datasets/02_cleaned_output.csv", index=False)
